Move legacy datapoint to class.
